Beginning of the implementation of the new Singleton pattern.
First, it makes all the Singletons use a unique and common pattern.

The new Singleton pattern switches who's in charge of the Singleton.
The instance itself manages itself, the generic ones are now more
like proxies with which the calls will eventually fall back on the
implementations.

Furthermore, we can now control the lifecycle of those Singletons.
They are instanciated whenever they are needed, registered by
themselves and destroyed in reverse order of their instanciations.