Fixed issue with code emition ordering with swig/lua
Lua version of the functions would take have equal precedence
over the "native" ones, causing issues when the declaration
order of the functions was lua first, native second; the lua
accepting functions were always called, even when they were
to be called with a swig userdata.

Well, that's how I understood it, if question arises, please
ask @twistedoe, he's the wizard that dit it.

This commit also reverts the last commit.